Update test_engine.py
BUG RISK

Comparisons to the singleton objects, like True, False, and None, should be done with identity, not equality. Use is or is not.
Identity checks are faster than equality checks. Also, the equality checks can result in unintended behaviour in some cases.

Exceptions:
Equality comparisons might be required in some cases, especially in SQLAlchemy, SQLObject, ORMs, etc. In these cases, it is safe to ignore this issue.